OXDEV-3103 deprecate AuthorizationService
